add plugins folder to config
