+++ lua/custom/plugins/filetree.lua
@@ -1,0 +1,13 @@
+return {
+  'nvim-neo-tree/neo-tree.nvim',
+  version = '*',
+  dependencies = {
+    'nvim-lua/plenary.nvim',
+    'nvim-tree/nvim-web-devicons', -- not strictly required, but recommended
+    'MunifTanjim/nui.nvim',
+  },
+  config = function()
+    require('neo-tree').setup {}
+    vim.keymap.set('n', '<leader>nt', '<cmd>Neotree toggle<cr>')
+  end,
+}
